Let ZMQ's bind decide who becomes the runtime, not a lock of our own
Replaces the attach-or-start probe and its node-wide file lock with the
mechanism that was already there: binding the local server port is a
kernel-level atomic claim, so of N processes racing to start (mpirun -np 4)
exactly one can win it. ClioInitImpl now just tries ServerInit and falls
back to ClientInit, and all three behaviours #1015 asks for fall out of that
one path -- including the foreign-program case, where the bind fails AND
ClientInit finds nobody answering, so init fails. Net -872/+57 lines, and
runtime_probe.{h,cc} plus its test are gone.

The fallback is not sufficient on its own, which is why the port claim moves
to the top of IpcManager::ServerInit. Left where it was -- last -- a loser
runs the whole of ServerInit before failing, and reaching the end sets
is_initialized_, which makes the subsequent ClientInit an early-return that
reports success. Measured: with a foreign program on the port, every rank
reported a healthy client attached to a runtime that did not exist.

Placement is load-bearing at both ends, and both ends cost a test run to
find:
  - after the chi_cur_worker_key_ TLS setup, because binding takes a lock and
    lock acquisition reads that key. Claiming earlier dereferences a garbage
    pointer and the daemon dies on startup (7 shutdown battletests).
  - after ServerInitShm, because the transport needs the shared-memory
    allocators. Claiming earlier left the whole data plane broken on the
    node whose runtime this was -- PutBlob rc=11 -- while the election
    itself looked perfectly healthy.
  - and still before WriteRuntimePidRecord and before is_initialized_, so a
    loser neither steals the live runtime's pid record nor short-circuits
    its own ClientInit.

Verified: 8-rank 2-node docker suite green (one owner per node, all
PutBlob/GetBlob); 4-rank race green; a foreign listener on the port makes
CLIO_INIT return false instead of silently succeeding.
