+++ context-runtime/src/ipc_manager.cc
@@ -67,7 +67,6 @@
 #include "clio_runtime/local_task_archives.h"
 #include "clio_runtime/pool_manager.h"
 #include "clio_runtime/runtime_pid_record.h"
-#include "clio_runtime/runtime_probe.h"
 #include "clio_runtime/scheduler/scheduler_factory.h"
 #include "clio_runtime/task_archives.h"
 
@@ -506,6 +505,31 @@ bool IpcManager::ServerInit() {
   if (!ServerInitShm()) {
     return false;
   }
+
+  // Claim the local server port before ANY state is created (issue #1015).
+  // Binding it is a kernel-level atomic claim, which makes it the natural
+  // mutual exclusion between processes racing to become this node's runtime:
+  // exactly one can win, and the losers fall back to attaching as clients.
+  //
+  // Placement is load-bearing at BOTH ends. It must come after the
+  // chi_cur_worker_key_ setup above -- binding takes a lock, and lock
+  // acquisition reads that TLS key, so claiming the port any earlier
+  // dereferences a garbage pointer and kills the daemon on startup. And it must
+  // come before ClearUserIpcs and everything after it: those steps are
+  // non-transactional with no rollback, and reaching the end of this function
+  // sets is_initialized_, which makes a later ClientInit a no-op that reports
+  // success -- a loser that got that far would report a healthy client attached
+  // to a runtime that does not exist.
+  //
+  // Nothing polls local_transport_; it exists to hold the port. So claiming it
+  // here costs nothing and does not expose a half-built runtime to clients.
+  if (!StartLocalServer()) {
+    HLOG(kInfo,
+         "IpcManager::ServerInit: local server port is already bound - "
+         "this process will not be the runtime");
+    return false;
+  }
+
 
   // Publish this runtime's pid as soon as its segments exist, and withdraw it
   // in UnlinkOwnArtifacts when they go: the record's lifetime brackets the
@@ -3072,14 +3096,6 @@ size_t IpcManager::ClearUserIpcs() {
 
   for (const auto &name : ctp::SystemInfo::ListDirectory(memfd_dir)) {
     std::string full_path = memfd_dir + "/" + name;
-
-    // Never reap a start lock (issue #1015). We are running INSIDE the locked
-    // region — unlinking the lock would let the next starter create a fresh
-    // inode and flock that instead, which excludes nobody, and two runtimes
-    // would come up on the same port after all.
-    if (name.rfind(kRuntimeStartLockPrefix, 0) == 0) {
-      continue;
-    }
 
     // The per-user memfd dir is shared by every runtime on this node (the
     // fallback topology runs several). Only reap leftovers from DEAD processes
