+++ context-runtime/util/clio_run_local_runtime.h
@@ -49,27 +49,19 @@
 
 #ifdef __APPLE__
 // macOS has no /proc: process introspection goes through libproc
-// (proc_pidpath).
+// (proc_pidpath) and the KERN_PROC sysctl (process state / zombie check).
 #include <libproc.h>
+#include <sys/proc.h>
+#include <sys/sysctl.h>
 #endif
 
 #include "clio_ctp/introspect/system_info.h"
 #include "clio_ctp/util/config_parse.h"
 #include "clio_runtime/config_manager.h"
 #include "clio_runtime/runtime_pid_record.h"
-#include "clio_runtime/runtime_probe.h"
 #include "clio_runtime/types.h"
 
 namespace clio::run::cli {
-
-// Segment/pid discovery and the pid-liveness tests are shared with the
-// attach-or-start probe the runtime itself uses (clio_runtime/runtime_probe.h)
-// — the stop CLI and CLIO_WITH_RUNTIME=1 must agree on what "a runtime is
-// running here" means, so there is exactly one implementation.
-using clio::run::DiscoverRuntimePid;
-using clio::run::MainSegmentPath;
-using clio::run::PidIsRunning;
-using clio::run::PidIsZombie;
 
 /**
  * Resolve the local runtime's port without requiring a live runtime:
@@ -87,6 +79,103 @@ inline u32 ResolveRuntimePort() {
     if (port > 0) return static_cast<u32>(port);
   }
   return 9413;
+}
+
+/**
+ * Path of the runtime's main shared-memory segment entry in the per-user
+ * memfd directory. Mirrors ConfigManager::GetSharedMemorySegmentName's
+ * default port-keyed recipe so it also works when no config is loaded.
+ * @param port the runtime port the segment name is keyed on
+ * @return absolute path of the main segment entry
+ */
+inline std::string MainSegmentPath(u32 port) {
+  const std::string name =
+      ctp::ConfigParse::ExpandPath("chi_main_segment_${USER}") + "_" +
+      std::to_string(port);
+  return ctp::SystemInfo::GetMemfdPath(name);
+}
+
+/**
+ * Discover the local runtime's pid: on Linux from its main segment symlink,
+ * whose target is /proc/<pid>/fd/<n> (the memfd the runtime created), and
+ * otherwise from the pid record the runtime writes alongside its segments
+ * (see clio_runtime/runtime_pid_record.h). POSIX-only;
+ * returns -1 on Windows.
+ * @param port the runtime port (segment names are port-keyed)
+ * @return the owning pid (may be dead), or -1 if the runtime left no trace
+ */
+inline int DiscoverRuntimePid(u32 port) {
+#ifdef _WIN32
+  (void)port;
+  return -1;
+#else
+  std::error_code ec;
+  auto target = std::filesystem::read_symlink(MainSegmentPath(port), ec);
+  if (!ec) {
+    const std::string t = target.string();
+    constexpr const char *kProc = "/proc/";
+    if (t.rfind(kProc, 0) == 0) {
+      int pid = std::atoi(t.c_str() + std::string(kProc).size());
+      if (pid > 0) {
+        return pid;
+      }
+    }
+  }
+  return ReadRuntimePidRecord(port);
+#endif
+}
+
+/**
+ * Check whether a pid is a zombie (exited but not yet reaped by its parent).
+ * kill(pid, 0) treats zombies as alive, but a zombie cannot run, service
+ * tasks, or hold sockets — for stop purposes it is dead. The stop CLI is a
+ * sibling of the runtime (the parent is Jarvis / a test harness / a shell),
+ * so it can never reap the zombie itself.
+ *
+ * This matters most on the kill-escalation path: after SIGKILL the runtime
+ * lingers as a zombie until its parent (the harness / Jarvis) reaps it, and
+ * a stop that cannot see through that reports "survived SIGKILL".
+ * @param pid candidate pid
+ * @return true if the process is a zombie
+ */
+inline bool PidIsZombie(int pid) {
+#if defined(__APPLE__)
+  // No /proc: ask the kernel for the process's state directly.
+  struct kinfo_proc info;
+  size_t len = sizeof(info);
+  int mib[4] = {CTL_KERN, KERN_PROC, KERN_PROC_PID, pid};
+  if (sysctl(mib, 4, &info, &len, nullptr, 0) != 0 || len == 0) {
+    return false;
+  }
+  return info.kp_proc.p_stat == SZOMB;
+#elif defined(__linux__)
+  std::ifstream stat_file("/proc/" + std::to_string(pid) + "/stat");
+  if (!stat_file.is_open()) {
+    return false;
+  }
+  std::string stat_line;
+  std::getline(stat_file, stat_line);
+  // Field 3 (state) follows the parenthesized comm, which may contain spaces.
+  size_t close_paren = stat_line.rfind(')');
+  if (close_paren == std::string::npos ||
+      close_paren + 2 >= stat_line.size()) {
+    return false;
+  }
+  return stat_line[close_paren + 2] == 'Z';
+#else
+  (void)pid;
+  return false;
+#endif
+}
+
+/**
+ * Check whether a pid is alive AND runnable (not a zombie). This is the
+ * liveness test the stop flow uses: a zombie counts as stopped.
+ * @param pid candidate pid
+ * @return true if the process exists and is not a zombie
+ */
+inline bool PidIsRunning(int pid) {
+  return ctp::SystemInfo::IsProcessAlive(pid) && !PidIsZombie(pid);
 }
 
 /**
